Add minimalist C shared library crate.
This shared object provides a C API/ABI for a dynamically linked library.

It is deliberately minimal, and can be linked against with the
included fips203.h, and subsequently linked against the .so built by
this cdylib.

The caller handles only serialized forms of the underlying cryptographic
objects, and is responsible for all memory allocation/destruction.

A more sophisticated library interface seems possible, but this one is
clean and can be easily bound to other programming languages that already
have the ability to wrap a C FFI.

This includes a simple test, which should be able to be run with:

    (cd tests && make)

On any system that has GNU make and a C compiler available.

I don't know how to integrate such a test into cargo's test suite,
though.