broaden criterion dependency range
This crate builds and tests fine against criterion 0.5.1 or 0.4.0;
while those versions are semver-incompatible, the crate itself uses a
common set of compatible features from criterion.

This adjustment in the explicit dependencies represents that
flexibility.

If criterion 0.5.1 is selected, and that package has a higher MSRV
than 0.72, then the aggregate MSRV will be higher.  But this package
won't be the limiting factor.