Ensure that the tests derive the SONAME correctly
When doing an initial build against a library with an SONAME, the
resulting ELF object stores the SONAME, not the literal filename that
was found for linking.

To make the test run correctly, we symlink the SONAME to the resultant
object.

This is a bit janky, for at least two reasons.

- it would be better to have the shared object be named with the
SONAME in the first place -- if we want to, then we can symlink to it
without the SONAME.

- if we have to find the SONAME in ffi/tests/Makefile, we should
really be able to derive the SONAME from the info in ffi/Cargo.toml
instead of hard-coding it in ffi/tests/Makefile