Clean up python packaging
This shuffling of metadata about the Python module will help submit
the package to PyPI.

Note that README.md contains the same content as the module-level
docstring.  I couldn't see a way to easily derive one from the other
so for the moment they should be kept manually in sync.

Dynamic metadata documentation only permits me to identify a file or
set of files, but not to extract an attr:

    https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html#dynamic-metadata

I also couldn't find a way to cleanly/explicitly represent the
dual-licensing in the Python classifiers, because Apache 2.0 is not
listed:

    https://pypi.org/classifiers/

Nonetheless, this is better than no documentation.